Move `Vec::from_par_iter` to the collect module
By moving it next to `special_collect_into`, we can keep that kludge
truly private, rather than `pub` and hidden.
