+++ src/par_iter/collect/mod.rs
@@ -1,4 +1,6 @@
-use super::{ParallelIterator, ExactParallelIterator};
+use super::{ParallelIterator, ExactParallelIterator, IntoParallelIterator};
+use super::from_par_iter::FromParallelIterator;
+use std::collections::LinkedList;
 use std::isize;
 use std::sync::atomic::{AtomicUsize, Ordering};
 
@@ -25,8 +27,7 @@ pub fn collect_into<PAR_ITER, T>(mut pi: PAR_ITER, v: &mut Vec<T>)
 /// *any* `ParallelIterator` here, and `CollectConsumer` has to also implement
 /// `UnindexedConsumer`.  That implementation panics `unreachable!` in case
 /// there's a bug where we actually do try to use this unindexed.
-#[doc(hidden)]
-pub fn special_collect_into<PAR_ITER, T>(pi: PAR_ITER, len: usize, v: &mut Vec<T>)
+fn special_collect_into<PAR_ITER, T>(pi: PAR_ITER, len: usize, v: &mut Vec<T>)
     where PAR_ITER: ParallelIterator<Item = T>,
           T: Send
 {
@@ -56,5 +57,43 @@ pub fn special_collect_into<PAR_ITER, T>(pi: PAR_ITER, len: usize, v: &mut Vec<T
                 len,
                 actual_writes);
         v.set_len(len);
+    }
+}
+
+/// Collect items from a parallel iterator into a freshly allocated vector.
+impl<T> FromParallelIterator<T> for Vec<T>
+    where T: Send
+{
+    fn from_par_iter<PAR_ITER>(par_iter: PAR_ITER) -> Self
+        where PAR_ITER: IntoParallelIterator<Item = T>
+    {
+        // See the vec_collect benchmarks in rayon-demo for different strategies.
+        let mut par_iter = par_iter.into_par_iter();
+        match par_iter.opt_len() {
+            Some(len) => {
+                // When Rust gets specialization, call `par_iter.collect_into()`
+                // for exact iterators.  Until then, `special_collect_into()` fakes
+                // the same thing on the promise that `opt_len()` is accurate.
+                let mut vec = vec![];
+                super::collect::special_collect_into(par_iter, len, &mut vec);
+                vec
+            }
+            None => {
+                // This works like `combine`, but `Vec::append` is more efficient than `extend`.
+                let list: LinkedList<_> = par_iter.fold(Vec::new, |mut vec, elem| {
+                        vec.push(elem);
+                        vec
+                    })
+                    .collect();
+
+                let len = list.iter().map(Vec::len).sum();
+                let start = Vec::with_capacity(len);
+                list.into_iter()
+                    .fold(start, |mut vec, mut sub| {
+                        vec.append(&mut sub);
+                        vec
+                    })
+            }
+        }
     }
 }
